Added the notification center class to Bump.
The NotificationCenter is a notification system that allows you to send notifications abstractly when events occur in your application. For example, if you're work on an event system, and your event completes, sometimes it would be nice to notify multiple parts of your application that that event was completed. The bump::NotificationCenter makes this type of notification very easy to do.