Added a patch to the FileInfo::exists() method that returns true if the path passes the symlink check.
Seems to be a bug in boost that at times, the symlink will actually come back as though it doesn't exist. However, the symlink check will return that the symlink is valid, meaning the symlink exists. This patch adds a second check to the exists method also looking for symlinks.

Ideally, this patch would not be necessary, but it appears that boost has a small bug.