Finished up the FileSystem unit test suite.
Not only is the FileSystem class now fully covered, but many methods were modified to no longer throw exceptions. The reasoning behind this was if a method already returns whether it succeeded or not, then the user should rely on that solely. Otherwise the user has to check whether an exception was thrown in addition to returning true/false.