Added a BUILD_DOCUMENTATION flag to the CMake build system allowing Doxygen to automatically run.
This is a really nice addition that allows a user to generate Doxygen files using CMake. The $BUMP_HOME/doc/doxyfile.cmake contains all the necessary flags for generating the doc files with all the proper relative paths. Then if the user has Doxygen installed, it will be picked up and a build target is created in the makefile or IDE project.