Fixed all the issues on Windows for the FileInfo and FileSystem classes, test suites and examples.
This modification is so large that it cannot be broken up into chunks. The major changes are the way symlinks are created. There are now different methods for creating a file symlink and a directory symlink. Additionally, all the methods in both classes that ask for a path object, the methods return path objects in the form of unix filepaths. This gives us consistency between unix and win operating systems. Each class, test suite and example had to be broken up into platform dependent sections in order to get everything working properly.