Moved the WIN32_USE_MP flag to Bump_WIN32_USE_MP and also set some strict boost restrictions to help CMake.
On Windows, you now have to use a static, multi-threaded boost library with a shared runtime to compile Bump. This allows CMake to properly find Boost. Otherwise on Windows, CMake keeps deleting the Boost library paths every time you hit configure. This is incredibly frustrating. Until this restriction on Boost becomes a problem, we'll leave it.