The Observer, KeyObserver and ObjectObserver classes are no longer using the BUMP_EXPORT preprocessor directive.
This causes issues on Windows with the declspec dllimports since the implementation of the templated classes is actually inside the header. Therefore, if we use the BUMP_EXPORT directive, the linker gets confused when linking other applications against the bump.lib. It can't find the symbols because the BUMP_EXPORT directive is telling the linker to find them in the bump.lib but they aren't in there. This patch fixes up the problem. The solution is to simply remove the BUMP_EXPORT directive altogether when the entire implementation is in the header.