Updated TextFileReader to return StringLists instead of boost pointers. Need to test if this is faster.
Changed a bunch of formatting in TextFileReader so that it matches the formatting in the rest of bump. Also removed the boost shared pointers int he code. Now we are returning bump objects like StringLists and Strings. This will need to be tested in the future to see which way is faster. This way is simpler if nothing else.