Unit tests for TextFileReaderTest testFooter and modifications to TextFileReader footer method.
Added unit tests for TextFileReaderTest testFooter method to test normal requests, negative numbers, floats, and requesting too large of a file.

Modifications to the TextFileReader::footer method now provide appropriate counting so we get the lines we want. The previous implementation moved the lines requested up by one so if you were looking for lines 8 and 9 it would return lines 7 and 8.